==> notebooks/1.0-tte-lda-sklearn.ipynb <==
# Clustering: LDA w/ sklearn
Prototyping NB for Clustering EUvsVirus Hackathon projects based on LDA using sklearn.

import os, warnings
warnings.simplefilter(action="ignore", category=FutureWarning)

import numpy as np
import pandas as pd

import spacy
from spacymoji import Emoji
import regex as re
import emoji

from sklearn.model_selection import GridSearchCV
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.decomposition import LatentDirichletAllocation
from joblib import dump, load

import pyLDAvis
import pyLDAvis.sklearn
pyLDAvis.enable_notebook()

%matplotlib inline

random_state = 23

## Load

Data

data_path = "/mnt/DATA/RWData Dropbox/EUvsVirus"
data_fpath = os.path.join(data_path, "all_data.tsv")

data = pd.read_csv(data_fpath, sep="\t")
print(data.info())
data.shape

NLP Model

nlp = spacy.load("en_core_web_md", disable=["ner", "parser", "tagger"])
emoji = Emoji(nlp, merge_spans=False)
nlp.add_pipe(emoji, first=True)

## Preprocessing

data.text.replace("", np.nan, inplace=True)
nan_removed = data.dropna(subset=["text"])
data_cleaned = nan_removed.drop_duplicates(subset=["text", "title"], keep="first")
raw_corpus = data_cleaned.text

raw_corpus.shape

### Text  Cleansing

Maybe later: Remove header texts; Split and structure text parts

def spacy_tokenizer(doc):
    tokens = [token for token in nlp(doc)]
    cleaned_tokens = []
                                        
    for token in tokens:
        if not any(map(lambda x: getattr(token, x), ["is_punct", "is_digit",
                                                     "is_space", "like_url",
                                                     "like_num", "like_email",
                                                     "is_stop"
                                                     ])):
            cleaned_tokens.append(token)
    
    # Post-processing steps
    lemmas = [token.lemma_ for token in cleaned_tokens]
    punct_removed = [re.sub(r"[.!?+:/\\-]", "", lemma) for lemma in lemmas]
    num_removed = [token for token in punct_removed if token.isalpha()]
    single_char_removed = [token for token in num_removed if len(token) > 1]
    
    return single_char_removed

## Topic Modeling
TFIDF + LDA

### Vectorizer

tfidf = TfidfVectorizer(lowercase=True, tokenizer=spacy_tokenizer, max_features=20000)
# total tokens 25744 -> get rid of infrequent
tfidf_v = tfidf.fit_transform(raw_corpus)

tfidf_v.shape

#tfidf.get_feature_names()

### Model

models = {
    "lda_032": LatentDirichletAllocation(n_components=32, n_jobs=1, random_state=random_state, verbose=1), # num subchallenges
    "lda_050": LatentDirichletAllocation(n_components=50, n_jobs=1, random_state=random_state, verbose=1), # best intrepret
    "lda_064": LatentDirichletAllocation(n_components=64, n_jobs=1, random_state=random_state, verbose=1), # 2*num subchallenges
    "lda_100": LatentDirichletAllocation(n_components=100, n_jobs=1, random_state=random_state, verbose=1)
}

model_persist_path = os.path.join("../models", "lda_sklearn")

for model in models:
    print("Fitting", model)
    models[model].fit(tfidf_v)
    print("Dump", model, "\n")
    dump(models[model], os.path.join(model_persist_path, model + ".joblib"))

def print_scores(model, name):
    """ Prints Log-Likelihood and Perplexity scores of model. """
    print("Scores for", name)
    print("Log Likelihood:", model.score(tfidf_v))
    print("Perplexity:", model.perplexity(tfidf_v), "\n")

for model in models:
    print_scores(models[model], model)

viz_path = os.path.join("../reports/figures/", "lda_sklearn")

for model in models:
        p = None
        p = pyLDAvis.sklearn.prepare(models[model], tfidf_v, tfidf, mds="tsne")
        pyLDAvis.save_html(p, os.path.join(viz_path, "topics_" + model + ".html"))
        print("topics_" + model + " persisted")

# Conclusion

Best interpretable model consist of 50 topics. Which might be too less for desired use-case. In general, model performance would benefit from incorporating more precise preprocessing steps. Due to lack of time in project, this will not happen (resp. will happen, but in R land).

**Not production ready**
