==> notebooks/3.0-ja-clustering2.ipynb <==
import pandas as pd
import numpy as np
import os
import datetime as dt
import re
#import gensim
import sklearn; # sklearn.get_config()['working_memory']
#from sklearn.metrics import pairwise_distances, pairwise_distances_chunked # not used
#from sklearn.metrics.pairwise import cosine_similarity, cosine_distances # not used
#from scipy.spatial.distance import pdist, cosine # not used

from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.cluster import AgglomerativeClustering
# Visualisation
import seaborn as sns
import plotly.express as px

#import nltk
from nltk.corpus import stopwords
from nltk.stem.porter import PorterStemmer
from nltk.stem.snowball import SnowballStemmer
from nltk.stem.lancaster import LancasterStemmer

data_dir = os.path.join(os.environ["HOME"],"Dropbox","EUvsVirus")

data_import = pd.read_csv(os.path.join(data_dir,"all_Data.tsv"), sep="\t")

stopwords = stopwords.words('english')

data = (data_import.dropna(subset=pd.Index(["text"]))
                   .assign(uniqueID = lambda x: x.ProjURL.str.extract(r"(?<=https://devpost.com/software/)(.+$)"),
                           text = lambda x: x.text.str.lower(), # not removing non-word chars yet because later we need the '\n' to separate on lines
                           ) 
)

Calculating the headings which should be removed

headings = (data.text.str.split('\n', expand=True) #splitting into lines
              .melt(value_name='lines') # melting into one column
                .lines.value_counts() # selecting the 'lines' column and counting frequency
                .reset_index()
)
headings.columns = ['headings','counts']


Visual inspection of the 200 most common lines showed that lines present 6+ times were not very defining. 

These will be removed by creating a variable called headings6plus and removing these headings from the text. Note: This is not removing the word from the entire text, just the line

headings.iloc[100:150]

headings6plus = headings.loc[headings.counts >= 6,'headings'] # counts of 6+
headings6plus

# lol = list of lines
# low = list of words
data = (data.assign(headings_rmvd = data.text.str.split('\n') # splitting into lines
                                             .apply(lambda lol: [line for line in lol if line not in list(headings6plus)]) # filtering out headings
                                             .apply(lambda lol: ' '.join(lol)), # joining the lines back together
                    words = lambda x: x.headings_rmvd.str.replace('\W',' ').str.split() # spliting on non-word chars and white space
                                                         .apply(lambda low: [word for word in low if word not in stopwords]) # removing stopwords
                   )
)



print(data.shape)
data.head()

Checking the word count of the submissions now.

45 submissions have only 1 word (after heading removal). These are not so useful.
After visual inspection I will remove entries with 10 words or less after headings removal. I don't think you could anyway get meaningful information from these submissions.

data.words.apply(len).value_counts().sort_index().head(15)

for x in data.text[data.words.apply(len) == 10]:
    print(x)
    print('---\n---\n')

data = data.loc[data.words.apply(len) > 10]

data

#from nltk.stem.porter import PorterStemmer
#from nltk.stem.snowball import SnowballStemmer
#from nltk.stem.lancaster import LancasterStemmer
ps = PorterStemmer()
ss = SnowballStemmer('english')
ls = LancasterStemmer()
# low = list of words
data = data.assign(stemmed_words = (data.words.apply(lambda low: [ps.stem(word) for word in low])
                                              #.apply(lambda low: [ss.stem(word) for word in low])
                                              #.apply(lambda low: [ls.stem(word) for word in low])
                                    )
                  )

data.stemmed_words

# low = list of words
tfidf_model = TfidfVectorizer(strip_accents='unicode', stop_words='english')
tfidf = tfidf_model.fit_transform(data.stemmed_words.apply(lambda low: ' '.join(low)))

There are way too many 'words' which are just numbers. These also need to be excluded.

tfidf_model.get_feature_names()[:10]

tfidf_vocab = pd.DataFrame({'words': tfidf_model.get_feature_names()})
print(len(tfidf_vocab))
tfidf_vocab[:1000]

tfidf_vocab = tfidf_vocab.assign(start_end_letter = lambda df: df.words.apply(lambda word: False if re.search('^[a-z].*[a-z]$',word) is None else True)
) # boolean for if the word starts and ends in a letter and contains zero or more characters between the first and last letter

tfidf_vocab.loc[tfidf_vocab.start_end_letter]

# low = list of words
tfidf_model = TfidfVectorizer(strip_accents='unicode', stop_words='english', vocabulary=tfidf_vocab.loc[tfidf_vocab.start_end_letter,'words'])
tfidf = tfidf_model.fit_transform(data.words.apply(lambda low: ' '.join(low)))

hc100 = AgglomerativeClustering(n_clusters=100,affinity = 'cosine', linkage='complete')
hc100.fit(tfidf.toarray())
#took approx 1min to run with sparse matrix of 1952 rows and 23084 words

px.histogram(x=hc100.labels_, nbins = 100)

hc200 = AgglomerativeClustering(n_clusters=200,affinity = 'cosine', linkage='complete')
hc200.fit(tfidf.toarray())
#took approx 2min to run with sparse matrix of 1952 rows and 23084 words

px.histogram(x=hc200.labels_, nbins = 200)

data = data.assign(hc100_labels = hc100.labels_,
                   hc200_labels = hc200.labels_)

x = data[['ProjURL','hc100_labels']]
x.columns = ['ID','cl']
x.to_csv(os.path.join(data_dir,'ja-clust100.csv'))

datax = data[['ProjURL','hc200_labels']]
x.columns = ['ID','cl']
x.to_csv(os.path.join(data_dir,'ja-clust200.csv'))

data_dir

def topwords(series,n):
    ''' returns the top_n words of a series as a list'''
    return series.value_counts().head(n).index.tolist()

split_words_and_labels.groupby('hc100_labels').word.agg(topwords, 5)

split_words_and_labels.groupby('hc200_labels').word.agg(topwords, 5)
